Make regist slightly more organized
Slight tweak to file structure, makes finding regist immediate in the
regist folder. All regist subordinate files have been moved to
_core/regist/inc/*
