+++ _core/regist/inc/addons.php
