+++ _core/regist/inc/fortune.php
