+++ _core/regist/inc/prune.php
