+++ _core/regist/inc/thumb/image.php
