+++ _core/regist/inc/thumb/thumb.php
