+++ _core/regist/inc/thumb/video.php
